requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.40.0
crispy-bootstrap5==2024.10
cryptography==44.0.1
dj-database-url==1.0.0
Django==5.2.2
django-allauth==65.4.1
django-axes==7.0.2
django-ckeditor-5==0.2.17
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-csp==4.0
django-formtools==2.5.1
django-js-asset==3.0.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-taggit==6.1.0
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.6
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
django-compressor


